Update code
Can not access SharedPreference file in application which with
directBootAware true.

Update box package

+++ lib-common/src/main/java/com/davidluoye/support/box/Longs.java
@@ -35,7 +35,7 @@ public class Longs {
         return Stream.of(array).mapToLong(it -> it != null ? it : 0).toArray();
     }
 
-    public static String toHexString(int value) {
+    public static String toHexString(long value) {
         return "0x" + Long.toHexString(value);
     }
 }
